Modify navigation menu items and add test dialog
Redundant or unnecessary items in the navigation menu were removed and replaced by items we'll actually need.

A test dialog was added for the GPS Location nav menu item, which will be needed for manually inputting a GPS address.  This was done by adding the GPSLocationDialogFragment file which serves as a wrapper for the dialog.