Add dummy flight vars dialog, add globals
A flight variables dialog was added for the nav drawer option, a text box is in place now but a slider may make more sense for adjusting the velocity and altitude of the drone.

Rudimentary support for globals has been added that can serve as a base for further usage.  To pass data back and forth for a Dialog you use an interface, which is defined in the Dialog file (here FlightVarsDialogFragment.java) and implemented in the activity you want to pass the data to (here MainActivity.java).

Other options for storing and modifying globals are to put them in the Application object, which can be prone to crashing, or storing them on disk and reading/writing as necessary.  My feeling is that interfaces would be simpler to implement than disk storage, so I went with this option- there isn't much data that needs to be passed anyway, at least not as of right now.

More information here:
http://www.developerphil.com/dont-store-data-in-the-application-object/