send velocity to drone, fix button state on exit
Velocity is sent properly, but next up is hardcoding in a minimum of 1 m/s, because the default speed (10 m/s?) is used if it's set to 0.

The app should no longer re-open with the button in the loading state, and loading state will now be preserved when the Settings menu is opened.